Refactor: Add class create_book
